% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_files.R
\name{clean_dataset_names}
\alias{clean_dataset_names}
\title{Clean file names of ENIGH data sets}
\usage{
clean_dataset_names(vector)
}
\arguments{
\item{vector}{A vector of file names.}
}
\value{
A vector of clean file names.
}
\description{
Helper function to clean ENIGH data.
}
