% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_menu}
\alias{data_menu}
\title{INEGI Data Menu}
\format{
\subsection{\code{data_menu}}{

A data frame with 1 rows and 3 columns:
\describe{
\item{Encuesta}{Name of the survey or data project.}
\item{Serie}{Series under which data is published.}
\item{Año}{Year span of data covered.}
...
}
}
}
\source{
\url{https://www.inegi.org.mx/}
}
\usage{
data_menu
}
\description{
A list of \href{https://www.inegi.org.mx/}{INEGI} data projects supported by \code{tidyinegi}.
}
\keyword{datasets}
