% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{get_enigh_var_labels}
\alias{get_enigh_var_labels}
\title{Get variable labels for ENIGH data sets}
\usage{
get_enigh_var_labels(data_set)
}
\arguments{
\item{data_set}{Name of the data set to get the labels from. The name must
match the name of the data set in the \href{https://www.inegi.org.mx/rnm/index.php/catalog/901/data-dictionary/}{INEGI website}.}
}
\value{
A tibble with the variable names and labels
}
\description{
Get variable labels for ENIGH data sets
}
\examples{
get_enigh_var_labels("viviendas")
get_enigh_var_labels("hogares")
}
