% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{handle_dichotomic}
\alias{handle_dichotomic}
\title{Adds factor labels to logical vectors in ENIGH data sets}
\usage{
handle_dichotomic(data)
}
\arguments{
\item{data}{An ENIGH data set in tibble format.}
}
\value{
A tibble with labelled logical variables.
}
\description{
Adds factor labels to logical vectors in ENIGH data sets
}
