% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{handle_factor_values}
\alias{handle_factor_values}
\title{Make factor variables in ENIGH data sets compatible with official
documentation.}
\usage{
handle_factor_values(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector with factor values compatible with official documentation.
}
\description{
Make factor variables in ENIGH data sets compatible with official
documentation.
}
\details{
Raw ENIGH data has anomallies in factor variables, such as
factor values that do not exactly match the official documentation.
}
\examples{
handle_factor_values(c("1", "2", "3"))
handle_factor_values(c("1", "2", "3", "a"))
handle_factor_values(c("01", "02", "03"))
}
