% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{handle_ignores}
\alias{handle_ignores}
\title{Adds factor labels to ignore type variables in ENIGH data sets}
\usage{
handle_ignores(data)
}
\arguments{
\item{data}{An ENIGH data set in tibble format.}
}
\value{
A tibble with labelled ignore type variables.
}
\description{
Adds factor labels to ignore type variables in ENIGH data sets
}
