% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{handle_numeric}
\alias{handle_numeric}
\title{Turn ENIGH numeric variables into numeric class}
\usage{
handle_numeric(data)
}
\arguments{
\item{data}{An ENIGH data set in tibble format.}
}
\value{
An ENIGH data set with numeric classes
}
\description{
Turn ENIGH numeric variables into numeric class
}
