% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{handle_single_values}
\alias{handle_single_values}
\title{Turn ENIGH single valued variables into logical}
\usage{
handle_single_values(data)
}
\arguments{
\item{data}{An ENIGH data set in tibble format.}
}
\value{
An ENIGH data set with logical variables
}
\description{
Turn ENIGH single valued variables into logical
}
