% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{has_characters}
\alias{has_characters}
\title{Does a vector contain alphabetic characters?}
\usage{
has_characters(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
TRUE if \code{x} contains alphabetic characters
}
\description{
Does a vector contain alphabetic characters?
}
\examples{
has_characters(c("1", "2", "3"))
has_characters(c("1", "2", "3", "a"))
}
