% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_sets.R
\name{has_problems}
\alias{has_problems}
\title{Detect cleaning problems in a vector}
\usage{
has_problems(vector)
}
\arguments{
\item{vector}{Vector to analyse}
}
\value{
\code{TRUE} if the vector has no valid values.
}
\description{
Helper function to clean ENIGH data.
}
