% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi_menu.R
\name{inegi_menu}
\alias{inegi_menu}
\title{INEGI Data Menu}
\usage{
inegi_menu(.quiet = FALSE)
}
\arguments{
\item{.quiet}{Use to avoid printing list.}
}
\value{
a \code{tibble}.
}
\description{
Display a list of \href{https://www.inegi.org.mx/}{INEGI} data projects supported
by \code{tidyinegi}. If you wish to load the menu visibly into memory,
use \code{data(data_menu)}.
}
\examples{
  inegi_menu()
}
