% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{is_dichotomic}
\alias{is_dichotomic}
\title{Is an ENIGH variable dichotomic?}
\usage{
is_dichotomic(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if \code{x} is dichotomic
}
\description{
Is an ENIGH variable dichotomic?
}
\examples{
is_dichotomic(c("0", "1", "2"))
is_dichotomic(c("1", "2", "3"))
is_dichotomic(c("1", "2"))
}
