% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{is_ignores}
\alias{is_ignores}
\title{Is an ENIGH variable an ignore type?}
\usage{
is_ignores(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if \code{x} is an ignore type variable (values 1, 2, 9)
}
\description{
Is an ENIGH variable an ignore type?
}
\examples{
is_ignores(c("1", "2", "9"))
is_ignores(c("1", "2", "3"))
is_ignores(c("1", "2"))
}
