% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{is_numeric}
\alias{is_numeric}
\title{Is an ENIGH variable numeric}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if \code{x} is numeric
}
\description{
Is an ENIGH variable numeric
}
