% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{is_single_value}
\alias{is_single_value}
\title{Is an ENIGH variable single valued?}
\usage{
is_single_value(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if \code{x} holds only one value, other than \code{NA}.
}
\description{
Is an ENIGH variable single valued?
}
