% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_files.R
\name{list_data_sets}
\alias{list_data_sets}
\title{List ENIGH data sets within a year}
\usage{
list_data_sets(year)
}
\arguments{
\item{year}{The year of the data sets.}
}
\value{
A tibble with the path and name of each data set.
}
\description{
List ENIGH data sets within a year
}
