% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_sets.R
\name{load_pre_clean_data}
\alias{load_pre_clean_data}
\title{Load pre clean data from ENIGH}
\usage{
load_pre_clean_data(data_set, year)
}
\arguments{
\item{data_set}{The name of a data set in ENIGH.}

\item{year}{ENIGH year. Defaults to global \code{.year}}
}
\value{
A tibble loaded into \code{.GlobalEnv}.
}
\description{
Helper function to clean ENIGH data.
}
