% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{nas}
\alias{nas}
\title{Compute number of NA's in every variable of an ENIGH data set}
\usage{
nas(data, ...)
}
\arguments{
\item{data}{An ENIGH data set in tibble format.}

\item{...}{Additional arguments passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}
}
\value{
A long tibble with the number of NA's in every variable.
}
\description{
Compute number of NA's in every variable of an ENIGH data set
}
