% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_sets.R
\name{problem_vars}
\alias{problem_vars}
\title{Number of problematic variables in data}
\usage{
problem_vars(data)
}
\arguments{
\item{data}{A tibble to search for cleaning errors.}
}
\value{
Number of columns with cleaning errors
}
\description{
Helper function to clean ENIGH data.
}
