% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_inegi_data.R
\name{raw_inegi_data}
\alias{raw_inegi_data}
\title{Download and unzip data from INEGI}
\usage{
raw_inegi_data(url, output_dir = "data")
}
\arguments{
\item{url}{A character string with the URL of the data to download. Must end in \code{.zip}}

\item{output_dir}{A character string with the path to the directory where the data will be saved.}
}
\value{
A character string with the path to the directory where the data was saved.
}
\description{
Helper function to clean ENIGH data.
}
