% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelling.R
\name{read_enigh_raw}
\alias{read_enigh_raw}
\title{Read raw ENIGH data sets}
\usage{
read_enigh_raw(data_set, year)
}
\arguments{
\item{data_set}{Name of the data set to read. The name must match the name of
the data set in the \href{https://www.inegi.org.mx/rnm/index.php/catalog/901/data-dictionary/}{INEGI website}.}

\item{year}{ENIGH year. Defaults to the value of the global.}
}
\value{
A tibble with the raw data set
}
\description{
Read raw ENIGH data sets
}
\examples{
year <- "2022"
data_set <- "poblacion"
read_enigh_raw(data_set, year)
}
