% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_files.R
\name{read_inegi_csv}
\alias{read_inegi_csv}
\title{Read an ENIGH csv file}
\usage{
read_inegi_csv(path, ...)
}
\arguments{
\item{path}{The path to the csv file.}

\item{...}{Additional arguments passed to \code{read_csv()}.}
}
\value{
A tibble.
}
\description{
Read an ENIGH csv file
}
