% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyinegi-package.R
\docType{package}
\name{tidyinegi-package}
\alias{tidyinegi}
\alias{tidyinegi-package}
\title{tidyinegi: Produce Analysis-Ready Data From INEGI}
\description{
Helper functions to clean data shipped in [tidyenigh](https://github.com/estebandegetau/tidyenigh).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/estebandegetau/tidyinegi}
  \item Report bugs at \url{https://github.com/estebandegetau/tidyinegi/issues}
}

}
\author{
\strong{Maintainer}: Esteban Degetau \email{estebandegetau@gmail.com} (\href{https://orcid.org/0009-0004-4095-8819}{ORCID})

}
\keyword{internal}
