% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data.R
\name{use_enigh}
\alias{use_enigh}
\title{Create ENIGH package data}
\usage{
use_enigh(data, data_set, year)
}
\arguments{
\item{data}{The data to be saved}

\item{data_set}{The name of an ENIGH data set}

\item{year}{The year of the ENIGH survey}
}
\description{
Makes it easy to ship data in a package, replicating what
\code{\link[usethis:use_data]{usethis::use_data()}} does, but for ENIGH data.
}
