% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyinegi-package.R
\docType{package}
\name{tidyinegi-package}
\alias{tidyinegi}
\alias{tidyinegi-package}
\title{tidyinegi: Access Analysis-Ready Data From [INEGI]("https://www.inegi.org.mx/")}
\description{
An open-source R package that allows users to programatically load analysis-ready INEGI data sets into R and in other formats (.dta, .xlsx). Why analysis-ready? Frequent users of INEGI data sets ought to know how painful it is to get from raw data into workable data. This package aims to bridge that gap, providing transparent documentation and inviting users to contribute.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/estebandegetau/tidyinegi}
  \item Report bugs at \url{https://github.com/estebandegetau/tidyinegi/issues}
}

}
\author{
\strong{Maintainer}: Esteban Degetau \email{estebandegetau@gmail.com} (\href{https://orcid.org/0009-0004-4095-8819}{ORCID})

}
\keyword{internal}
